refactor: drop backward-compat shims for cache and naics strategies
- Delete sbir_etl/utils/cache/base_cache.py (only re-exported APICache
  as BaseDataFrameCache) and the alias line in api_cache.py. Rename
  tests/unit/utils/test_base_cache.py to test_api_cache_core.py and
  remove the now-tautological alias-test class.
- Delete four 4-line re-export shims in
  sbir_etl/enrichers/naics/fiscal/strategies/ (agency_defaults.py,
  original_data.py, sector_fallback.py, topic_code.py). Update test
  imports to pull directly from simple_strategies. Public API via
  strategies/__init__.py is unchanged.

Net -34 LOC. Public API preserved. All touched tests pass; pre-existing
HTML-report and plotly-dashboard failures verified unchanged on HEAD.
