--- sbir_etl/enrichers/naics/fiscal/strategies/agency_defaults.py
@@ -1,4 +1,0 @@
-"""Backward-compat re-export. Canonical location: simple_strategies.py"""
-from .simple_strategies import AgencyDefaultsStrategy
-
-__all__ = ["AgencyDefaultsStrategy"]
