--- sbir_etl/enrichers/naics/fiscal/strategies/original_data.py
@@ -1,4 +1,0 @@
-"""Backward-compat re-export. Canonical location: simple_strategies.py"""
-from .simple_strategies import OriginalDataStrategy
-
-__all__ = ["OriginalDataStrategy"]
