--- sbir_etl/enrichers/naics/fiscal/strategies/sector_fallback.py
@@ -1,4 +1,0 @@
-"""Backward-compat re-export. Canonical location: simple_strategies.py"""
-from .simple_strategies import SectorFallbackStrategy
-
-__all__ = ["SectorFallbackStrategy"]
