--- sbir_etl/enrichers/naics/fiscal/strategies/topic_code.py
@@ -1,4 +1,0 @@
-"""Backward-compat re-export. Canonical location: simple_strategies.py"""
-from .simple_strategies import TopicCodeStrategy
-
-__all__ = ["TopicCodeStrategy"]
