--- sbir_etl/utils/cache/base_cache.py
@@ -1,5 +1,0 @@
-"""Backward compatibility shim — BaseDataFrameCache now lives in api_cache.py."""
-
-from .api_cache import APICache as BaseDataFrameCache
-
-__all__ = ["BaseDataFrameCache"]
