+++ tests/unit/enrichers/naics/test_fiscal_strategies.py
@@ -11,7 +11,7 @@ class TestOriginalDataStrategy:
 
     def test_extracts_naics_from_standard_column(self):
         """Test extraction from NAICS column."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
 
         strategy = OriginalDataStrategy()
         row = pd.Series({"NAICS": "541715", "company_name": "Test Corp"})
@@ -26,7 +26,7 @@ class TestOriginalDataStrategy:
 
     def test_extracts_naics_from_alternative_columns(self):
         """Test extraction from alternative column names."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
 
         strategy = OriginalDataStrategy()
 
@@ -39,7 +39,7 @@ class TestOriginalDataStrategy:
 
     def test_returns_none_when_no_naics_column(self):
         """Test returns None when no NAICS column present."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
 
         strategy = OriginalDataStrategy()
         row = pd.Series({"company_name": "Test Corp", "award_amount": 100000})
@@ -50,7 +50,7 @@ class TestOriginalDataStrategy:
 
     def test_returns_none_for_null_naics(self):
         """Test returns None when NAICS value is null."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
 
         strategy = OriginalDataStrategy()
         row = pd.Series({"NAICS": None, "company_name": "Test Corp"})
@@ -61,7 +61,7 @@ class TestOriginalDataStrategy:
 
     def test_strategy_properties(self):
         """Test strategy name and confidence level."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
 
         strategy = OriginalDataStrategy()
 
@@ -74,7 +74,7 @@ class TestSectorFallbackStrategy:
 
     def test_returns_default_fallback_code(self):
         """Test returns default 5415 code."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.sector_fallback import SectorFallbackStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import SectorFallbackStrategy
 
         strategy = SectorFallbackStrategy()
         row = pd.Series({"company_name": "Test Corp"})
@@ -89,7 +89,7 @@ class TestSectorFallbackStrategy:
 
     def test_custom_fallback_code(self):
         """Test with custom fallback code."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.sector_fallback import SectorFallbackStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import SectorFallbackStrategy
 
         strategy = SectorFallbackStrategy(fallback_code="541712")
         row = pd.Series({"company_name": "Test Corp"})
@@ -101,7 +101,7 @@ class TestSectorFallbackStrategy:
 
     def test_always_succeeds(self):
         """Test fallback always returns a result."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.sector_fallback import SectorFallbackStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import SectorFallbackStrategy
 
         strategy = SectorFallbackStrategy()
 
@@ -114,7 +114,7 @@ class TestSectorFallbackStrategy:
 
     def test_strategy_properties(self):
         """Test strategy name and confidence level."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.sector_fallback import SectorFallbackStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import SectorFallbackStrategy
 
         strategy = SectorFallbackStrategy()
 
@@ -127,7 +127,7 @@ class TestAgencyDefaultsStrategy:
 
     def test_returns_dod_default(self):
         """Test DOD agency returns aerospace manufacturing code."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 
         strategy = AgencyDefaultsStrategy()
         row = pd.Series({"agency": "DOD", "company_name": "Test Corp"})
@@ -140,7 +140,7 @@ class TestAgencyDefaultsStrategy:
 
     def test_returns_hhs_default(self):
         """Test HHS agency returns biotech R&D code."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 
         strategy = AgencyDefaultsStrategy()
         row = pd.Series({"agency": "HHS", "company_name": "Test Corp"})
@@ -152,7 +152,7 @@ class TestAgencyDefaultsStrategy:
 
     def test_returns_none_for_unknown_agency(self):
         """Test returns None for unknown agency."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 
         strategy = AgencyDefaultsStrategy()
         row = pd.Series({"agency": "UNKNOWN_AGENCY", "company_name": "Test Corp"})
@@ -163,7 +163,7 @@ class TestAgencyDefaultsStrategy:
 
     def test_returns_none_when_no_agency(self):
         """Test returns None when agency column missing."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 
         strategy = AgencyDefaultsStrategy()
         row = pd.Series({"company_name": "Test Corp"})
@@ -174,7 +174,7 @@ class TestAgencyDefaultsStrategy:
 
     def test_strategy_properties(self):
         """Test strategy name and confidence level."""
-        from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+        from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 
         strategy = AgencyDefaultsStrategy()
