+++ tests/unit/enrichers/test_naics_strategies.py
@@ -16,12 +16,12 @@ from datetime import datetime
 import pandas as pd
 import pytest
 
-from sbir_etl.enrichers.naics.fiscal.strategies.agency_defaults import AgencyDefaultsStrategy
+from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import AgencyDefaultsStrategy
 from sbir_etl.enrichers.naics.fiscal.strategies.base import NAICSEnrichmentResult
-from sbir_etl.enrichers.naics.fiscal.strategies.original_data import OriginalDataStrategy
-from sbir_etl.enrichers.naics.fiscal.strategies.sector_fallback import SectorFallbackStrategy
+from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import OriginalDataStrategy
+from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import SectorFallbackStrategy
 from sbir_etl.enrichers.naics.fiscal.strategies.text_inference import TextInferenceStrategy
-from sbir_etl.enrichers.naics.fiscal.strategies.topic_code import TopicCodeStrategy
+from sbir_etl.enrichers.naics.fiscal.strategies.simple_strategies import TopicCodeStrategy
 from sbir_etl.enrichers.naics.fiscal.strategies.usaspending_dataframe import USAspendingDataFrameStrategy
 from sbir_etl.enrichers.naics.fiscal.utils import normalize_naics_code, validate_naics_code
